Added common activities package and added a simple WalkActivity.
